+++ packages/common/authUtils/.eslintrc.js
@@ -1,5 +1,12 @@
 module.exports = {
   extends: ['@moralisweb3'],
   plugins: ['jest'],
-  ignorePatterns: ['**/lib/**/*', '**/*.test.ts', '**/dist/**/*', '**/build/**/*', '**/generated/**/*'],
+  ignorePatterns: [
+    '**/lib/**/*',
+    '**/*.test.ts',
+    '**/dist/**/*',
+    '**/build/**/*',
+    '**/generated/**/*',
+    'src/operations/openapi.ts',
+  ],
 };